almost done

+++ libft/ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: lpellier <lpellier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/30 11:14:26 by lpellier          #+#    #+#             */
-/*   Updated: 2020/01/03 16:43:53 by lpellier         ###   ########.fr       */
+/*   Updated: 2020/01/23 17:00:08 by lpellier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_intlen(int n)
+int		ft_intlen(long n)
 {
 	int		count;
